+++ src/model/model.h
@@ -23,11 +23,11 @@ public:
   /// /brief Adds task with description to todo list.
   /// /param task The task to add.
   /// /param path Path of task to add.
-  void add(Task task, const std::vector<u16> &path);
+  void add(Task task, const std::vector<u64> &path);
 
   /// /brief Remove task from the todo list.
   /// /param path Path of the task to remove.
-  void remove(const std::vector<u16> &path);
+  void remove(const std::vector<u64> &path);
 
   /// \brief Clears (empties) the list.
   void clear();
@@ -43,12 +43,12 @@ public:
   /// \brief Changes the completion status of a task.
   /// \param path Path of the task to change.
   /// \param status Status to change to.
-  void change_task_status(const std::vector<u16> &path, const Status status);
+  void change_task_status(const std::vector<u64> &path, const Status status);
 
   /// \brief Changes the completion status of a task.
   /// \param path Path of the task to change.
   /// \param prio Priority to change to.
-  void change_task_priority(const std::vector<u16> &path, const int priority);
+  void change_task_priority(const std::vector<u64> &path, const int priority);
 
   /// \brief Returns a const reference of the todo list vector.
   const std::vector<Task> &get_list();
